Geojson tests adapted to MVT
